+++ src/mkhelp.pl
@@ -23,13 +23,6 @@
 #
 ###########################################################################
 
-# Yeah, I know, probably 1000 other persons already wrote a script like
-# this, but I'll tell ya:
-
-# THEY DON'T FIT ME :-)
-
-# Get readme file as parameter:
-
 if($ARGV[0] eq "-c") {
     $c=1;
     shift @ARGV;
@@ -53,6 +46,8 @@ print <<HEAD
  */
 #ifdef USE_MANUAL
 #include "tool_hugehelp.h"
+#include "tool_help.h"
+
 HEAD
     ;
 if($c) {
@@ -111,23 +106,25 @@ static void zfree_func(voidpf opaque, voidpf ptr)
   (void) opaque;
   free(ptr);
 }
+
+#define HEADERLEN 10
+
 /* Decompress and send to stdout a gzip-compressed buffer */
 void hugehelp(void)
 {
   unsigned char *buf;
-  int status, headerlen;
+  int status;
   z_stream z;
 
   /* Make sure no gzip options are set */
   if(hugehelpgz[3] & 0xfe)
     return;
 
-  headerlen = 10;
   memset(&z, 0, sizeof(z_stream));
   z.zalloc = (alloc_func)zalloc_func;
   z.zfree = (free_func)zfree_func;
-  z.avail_in = (unsigned int)(sizeof(hugehelpgz) - headerlen);
-  z.next_in = (unsigned char *)hugehelpgz + headerlen;
+  z.avail_in = (unsigned int)(sizeof(hugehelpgz) - HEADERLEN);
+  z.next_in = (unsigned char *)hugehelpgz + HEADERLEN;
 
   if(inflateInit2(&z, -MAX_WBITS) != Z_OK)
     return;
@@ -144,7 +141,49 @@ void hugehelp(void)
           break;
       }
       else
-        break;    /* Error */
+        break;    /* error */
+    }
+    free(buf);
+  }
+  inflateEnd(&z);
+}
+/* Show the help text for the 'arg' curl argument on stdout */
+void showhelp(const char *trigger, const char *arg, const char *endarg)
+{
+  unsigned char *buf;
+  int status;
+  z_stream z;
+  struct scan_ctx ctx;
+  inithelpscan(&ctx, trigger, arg, endarg);
+
+  /* Make sure no gzip options are set */
+  if(hugehelpgz[3] & 0xfe)
+    return;
+
+  memset(&z, 0, sizeof(z_stream));
+  z.zalloc = (alloc_func)zalloc_func;
+  z.zfree = (free_func)zfree_func;
+  z.avail_in = (unsigned int)(sizeof(hugehelpgz) - HEADERLEN);
+  z.next_in = (unsigned char *)hugehelpgz + HEADERLEN;
+
+  if(inflateInit2(&z, -MAX_WBITS) != Z_OK)
+    return;
+
+  buf = malloc(BUF_SIZE);
+  if(buf) {
+    while(1) {
+      z.avail_out = BUF_SIZE;
+      z.next_out = buf;
+      status = inflate(&z, Z_SYNC_FLUSH);
+      if(status == Z_OK || status == Z_STREAM_END) {
+        size_t len = BUF_SIZE - z.avail_out;
+        if(!helpscan(buf, len, &ctx))
+          break;
+        if(status == Z_STREAM_END)
+          break;
+      }
+      else
+        break;    /* error */
     }
     free(buf);
   }
@@ -187,6 +226,21 @@ void hugehelp(void)
   int i = 0;
   while(curlman[i])
     puts(curlman[i++]);
+}
+
+/* Show the help text for the 'arg' curl argument on stdout */
+void showhelp(const char *trigger, const char *arg, const char *endarg)
+{
+  int i = 0;
+  struct scan_ctx ctx;
+  inithelpscan(&ctx, trigger, arg, endarg);
+  while(curlman[i]) {
+    size_t len = strlen(curlman[i]);
+    if(!helpscan((unsigned char *)curlman[i], len, &ctx) ||
+       !helpscan((unsigned char *)"\\n", 1, &ctx))
+      break;
+    i++;
+  }
 }
 ENDLINE
     ;
