+++ tests/data/Makefile.am
@@ -221,7 +221,8 @@ test1670 test1671 \
 \
 test1680 test1681 test1682 test1683 \
 \
-test1700 test1701 test1702 test1703 test1704 test1705 test1706 \
+test1700 test1701 test1702 test1703 test1704 test1705 test1706 test1707 \
+test1708 test1709 test1710 \
 \
 test1800 test1801 \
 \
