+++ tests/data/test1461
@@ -13,6 +13,9 @@
 #
 # Client-side
 <client>
+<features>
+manual
+</features>
 <server>
 none
 </server>
@@ -34,7 +37,7 @@ curl important --help
 Usage: curl [options...] <url>
  -d, --data <data>           HTTP POST data
  -f, --fail                  Fail fast with no output on HTTP errors
- -h, --help <category>       Get help for commands
+ -h, --help <subject>        Get help for commands
  -o, --output <file>         Write to file instead of stdout
  -O, --remote-name           Write output to file named as remote file
  -i, --show-headers          Show response headers in output
@@ -50,7 +53,8 @@ Use "--help category" to get an overview of all categories, which are:
 auth, connection, curl, deprecated, dns, file, ftp, global, http, imap, ldap, 
 output, pop3, post, proxy, scp, sftp, smtp, ssh, telnet, tftp, timeout, tls, 
 upload, verbose.
-For all options use the manual or "--help all".
+Use "--help all" to list all options
+Use "--help [option]" to view documentation for a given option
 </stdout>
 </verify>
 </testcase>
